+++ src/declaration/network-port-number.ts
@@ -1,5 +1,11 @@
 import { createNetworkPortNumberConstraint } from "../constraint/network-port-number.js";
 import {
+  assertRangeSpec,
+  createRangeConstraint,
+  hasNumberRangeConstraint,
+  type RangeConstraintSpec,
+} from "../constraint/range.js";
+import {
   Declaration,
   Options as DeclarationOptions,
   Value,
@@ -7,11 +13,14 @@ import {
   type ExactOptions,
 } from "../declaration.js";
 import { registerVariable } from "../environment.js";
+import { normalize } from "../error.js";
 import { type Example } from "../example.js";
 import { resolve } from "../maybe.js";
 import { ScalarSchema, createScalar, toString } from "../schema.js";
+import { SpecError } from "../variable.js";
 
-export type Options = DeclarationOptions<number>;
+export type Options = DeclarationOptions<number> &
+  Partial<RangeConstraintSpec<number>>;
 
 export function networkPortNumber<O extends Options>(
   name: string,
@@ -20,7 +29,7 @@ export function networkPortNumber<O extends Options>(
 ): Declaration<number, O> {
   const { isSensitive = false } = options;
   const def = defaultFromOptions(options);
-  const schema = createSchema();
+  const schema = createSchema(name, options);
 
   const v = registerVariable({
     name,
@@ -38,7 +47,7 @@ export function networkPortNumber<O extends Options>(
   };
 }
 
-function createSchema(): ScalarSchema<number> {
+function createSchema(name: string, options: Options): ScalarSchema<number> {
   function unmarshal(v: string): number {
     if (!/^\d*$/.test(v)) throw new Error("must be an unsigned integer");
     if (v !== "0" && v.startsWith("0")) {
@@ -48,9 +57,18 @@ function createSchema(): ScalarSchema<number> {
     return Number(v);
   }
 
-  return createScalar("port number", toString, unmarshal, [
-    createNetworkPortNumberConstraint(),
-  ]);
+  const constraints = [createNetworkPortNumberConstraint()];
+
+  try {
+    if (hasNumberRangeConstraint(options)) {
+      assertRangeSpec(constraints, options);
+      constraints.push(createRangeConstraint(options));
+    }
+  } catch (error) {
+    throw new SpecError(name, normalize(error));
+  }
+
+  return createScalar("port number", toString, unmarshal, constraints);
 }
 
 function buildExamples(): Example[] {